+++ client/styles/GlobalStyle.js
@@ -29,7 +29,7 @@ const GlobalStyle = createGlobalStyle`
     top: 0;
     left: 0;
   }
-   
+
   .box {
     position: relative;
     width: 50%;
@@ -64,7 +64,7 @@ const GlobalStyle = createGlobalStyle`
 export const Button = styled(motion.button)`
   padding: 0.5rem;
   background: ${(props) =>
-    props.secondary ? 'var(--white)' : 'var(--orange)'};
+    props.secondary ? 'var(--yellow)' : 'var(--orange)'};
   border: 0;
   border-radius: 1rem;
   margin: 0.5rem;
