+++ client/styles/Splits.js
@@ -13,8 +13,17 @@ export const SplitRow = styled.div`
   padding: 0 1em;
 `;
 
+export const SplitTab = styled.div`
+  width: calc(100% - 4rem);
+  display: flex;
+  flex-flow: column nowrap;
+  padding: 1rem;
+  margin: 0 1rem;
+  background-color: var(--white);
+  color: var(--blue);
+`;
 export const SplitConfirm = styled.div`
   display: flex;
   font-weight: bold;
-  padding: 1em;
+  padding: 1rem;
 `;
