+++ server/app.js
@@ -28,7 +28,9 @@ app.get('/', async (req, res, next) => {
 });
 
 app.use((err, req, res, next) => {
+  console.log(err);
   res.status(err.status || 500).send(err);
+  next(err);
 });
 
 module.exports = app;
